+++ packages/perseus-editor/src/components/widget-editor.tsx
@@ -229,7 +229,6 @@ class WidgetEditor extends React.Component<
                         widgetInfo={widgetInfo}
                         onAlignmentChange={this._handleAlignmentChange}
                         isEditingDisabled={isEditingDisabled}
-                        apiOptions={this.props.apiOptions}
                     />
                 )}
                 <div