+++ packages/perseus/src/widgets/mock-widgets/mock-widget.tsx
@@ -87,14 +87,12 @@ class MockWidgetComponent extends React.Component<Props> implements Widget {
     render(): React.ReactNode {
         return (
             <View style={styles.widgetContainer}>
-                {/* eslint-disable-next-line jsx-a11y/no-redundant-roles -- TODO(LEMS-2871): Address a11y error */}
                 <TextField
                     ref={(ref) => (this.inputRef = ref)}
                     aria-label="Mock Widget"
                     value={this.props.userInput.currentValue}
                     onChange={this.handleChange}
                     id={this.props.widgetId}
-                    role="textbox"
                     onFocus={this.focusInputPath}
                     onBlur={this.blurInputPath}
                 />