+++ packages/perseus/src/widgets/numeric-input/utils.ts
@@ -25,6 +25,7 @@ const NumericExampleStrings: {
             : strings.simplifiedImproperExample,
     mixed: (form, strings) => strings.mixedExample,
     decimal: (form, strings) => strings.decimalExample,
+    percent: (form, strings) => strings.percentExample,
     pi: (form, strings) => strings.piExample,
 };