+++ packages/perseus/src/widgets/python-program/python-program.tsx
@@ -71,8 +71,8 @@ class PythonProgram extends React.Component<Props> implements Widget {
         // http://www.html5rocks.com/en/tutorials/security/sandboxed-iframes/
         return (
             <View style={styles.container}>
-                {/* eslint-disable-next-line jsx-a11y/iframe-has-title -- TODO(LEMS-2871): Address a11y error */}
                 <iframe
+                    title={this.context.strings.pythonProgram}
                     sandbox={sandboxOptions}
                     src={url}
                     style={iframeStyle}