+++ packages/perseus-linter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/perseus-linter
 
+## 5.0.12
+
+### Patch Changes
+
+-   Updated dependencies [[`c69e69a523`](https://github.com/Khan/perseus/commit/c69e69a523b942628b6c1cbe824c1521e878646f)]:
+    -   @khanacademy/perseus-core@27.0.3
+    -   @khanacademy/kmath@2.4.15
+
 ## 5.0.11
 
 ### Patch Changes